% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiR-package.R
\docType{package}
\name{aiR-package}
\alias{aiR}
\alias{aiR-package}
\title{aiR: Download Air Quality Data via the 'AirNow' API}
\description{
'AirNow' <https://www.airnow.gov> reports air quality information using an Air Quality Index (AQI). This package provides methods to query, retrieve, and interact these data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/briandconnelly/aiR}
  \item \url{https://docs.airnowapi.org/}
  \item Report bugs at \url{https://github.com/briandconnelly/aiR/issues}
}

}
\author{
\strong{Maintainer}: Brian Connelly \email{bconnelly@ookla.com} (\href{https://orcid.org/0000-0002-9948-0379}{ORCID})

}
\keyword{internal}
