% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{aiR}
\alias{aiR}
\title{aiR}
\usage{
aiR(data, input.col, var.classify, aiRnet, cycles = 100,
  sample.method = "Sample", sample.size = 0.5, batch.size = "all",
  sample.Factor = NULL, na.rm = TRUE, test.rate = "fast")
}
\arguments{
\item{data}{Data frame that contains all named columns needed}

\item{input.col}{character vector of column names for aiRnet}

\item{var.classify}{index or column name of vector that contains classifying values}

\item{aiRnet}{aiRnet object generated by aiRnet function}

\item{cycles}{Number of cycles done to correct.}

\item{sample.method}{Method to save internal subset of data as the training data. "Sample" to take
a random sample of all rows in data as training set. "Factor" to indicate if you are training
on rows containing a particular factor level.}

\item{sample.size}{Number between (0-1) that modifies how much of desired sample.method data is used.
Default set to 0.5. Excluded rows will be used as test examples and not affect aiRnet.}

\item{batch.size}{Number indicating how many rows to make batches from training sample. Default set to "all"
for no batches to be made.}

\item{sample.Factor}{Necessary when sample.method set to "Factor". Assign as vector of length 2 in the following form
c("column.index/name","factor.level"). Can assign only column name or index but first level is chosen in this case.}

\item{na.rm}{remove NAs, default set to TRUE. Function likely to fail with NAs}

\item{test.rate}{"none" or "fast" or "slow" or integer specifying how often the test model is checked. Test model is
used only when sample.size < 0.8 is set, otherwise the full model is used instead. "fast" indicates model is tested
at begining and end. "slow" indicates full model is tested each cycle. An integer indicates how many cycles pass
before testing the model. e.i. 1 is equivelent to "slow", 2 is every other cycle, 3 is every third cycle, ... etc.
"none" indicates that the model will not be verified on any model.}
}
\value{
data frame of loss and the aiRnet of the training values. data set used is optional
}
\description{
aiR
}
