% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_codes.R
\name{air_codes}
\alias{air_codes}
\title{Get filter codes to supply to filtering arguments}
\usage{
air_codes(arg = "institutions")
}
\arguments{
\item{code}{Which argument would you like to see filters for?}
}
\description{
Get filter codes to supply to filtering arguments
}
\examples{
air_codes("institions")))
}
