% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_vars.R
\name{air_vars}
\alias{air_vars}
\title{Elements of DCARS that air packages can use.}
\usage{
air_vars(cube)
}
\arguments{
\item{cube}{Return a list of possible measures and rows to query using AIR in that cube}
}
\description{
Elements of DCARS that air packages can use.
}
\examples{
air_vars("LERS")

}
