% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{aiRactivation}
\alias{aiRactivation}
\title{aiRactivation}
\usage{
aiRactivation(aiRnet, data)
}
\arguments{
\item{aiRnet}{aiRnet object to be trained}

\item{data}{datasets to pass through model}
}
\value{
returns a list of length(aiRnet)+1 where the first index is data. Subsequent indexes show
the activation levels of neurons per data observation. May be helpful in data visualization.
}
\description{
calculates and saves the activation of the network.
}
