% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{aiRclassify}
\alias{aiRclassify}
\title{aiRclassify}
\usage{
aiRclassify(data, factor, aiRnet)
}
\arguments{
\item{data}{data of input values}

\item{factor}{character vector of factors to classify. levels must be in same order as
levels presented in original data.}

\item{aiRnet}{aiRnet object}
}
\value{
returns a list of classification values for each
observation along with its activation in the last layer.
}
\description{
classifies each observation by taking the max value of the last layer.
}
