% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{aiRcost}
\alias{aiRcost}
\title{aiRcost}
\usage{
aiRcost(data, classify, class.levels)
}
\arguments{
\item{data}{takes in final activation level of a network.}

\item{classify}{Factor vector of correct classifications corresponding to data observations.}

\item{class.levels}{the levels of the classifier vector}
}
\value{
returns aiRcost object containing the following labels cost, dcost, obser.cost, total.cost, mean.cost
   cost: cost per value,
   dcost: differential cost per value,
   obser.cost: cost per observation,
   total.cost: batch total cost,
   mean.cost: batch mean cost
}
\description{
Calculates the resulting cost of the network for a given dataset activation.
}
