% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{aiRdevelop}
\alias{aiRdevelop}
\title{aiRdevelop}
\usage{
aiRdevelop(data, var.classify, aiRnet, train.method = "Sample",
  sample.size = 0.5, batch.size, train.Factor = NULL, steps = 15,
  range.size = 1000)
}
\arguments{
\item{data}{Data frame that contains all named columns needed}

\item{var.classify}{index or column name of vector that contains classifying values}

\item{aiRnet}{aiRnet object}

\item{train.method}{Method to save internal subset of data as the training data. "Sample" to take
a random sample of all rows in data as training set. "Factor" to indicate if you are training
on rows containing a particular factor level.}

\item{sample.size}{Number between (0-1) that modifies how much of desired train.method data is used.
Default set to 0.5. Excluded rows will be used as test examples and not affect aiRaiRnet.}

\item{batch.size}{Number indicating how many rows to make batches from training sample. Default set to "all"
for no batches to be made.}

\item{train.Factor}{Necessary when train.method set to "Factor". Assign as vector of length 2 in the following form
c("column.index/name","factor.level"). Can assign only column name or index but first level is chosen in this case.}

\item{steps}{Number of classification captures desired to view}

\item{range.size}{total amount of data points used to visualize input space.}

\item{na.rm}{remove NAs, default set to TRUE. Function likely to fail with NAs}
}
\value{
tidy data frame of the classification and max.node at each step
}
