% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{aiRfresh}
\alias{aiRfresh}
\title{aiRfresh}
\usage{
aiRfresh(aiRnet)
}
\arguments{
\item{aiRnet}{aiRnet object}
}
\value{
aiRnet object adjusted via change.w and change.b
}
\description{
applies the changes of change.w and change.b to the layer's respective
weights and biasis
}
