% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{aiRhidden}
\alias{aiRhidden}
\title{aiRhidden}
\usage{
aiRhidden(...)
}
\arguments{
\item{...}{takes aiRnet objects as inputs and creats a list as aiRhidden}
}
\value{
generates aiRhidden from aiRnet objects passed
}
\description{
Builds a network of aiRlayers objects with random weight and bias as a list
}
