% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation.R
\name{aiRlayer}
\alias{aiRlayer}
\title{aiRlayer}
\usage{
aiRlayer(dim1, dim2)
}
\arguments{
\item{dim1}{row dimensions}

\item{dim2}{column dimensions}
}
\value{
generates aiRlayer object of specified dimensions
}
\description{
aiRlayer
}
