% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{aiRlayer_trans.aiRlayer}
\alias{aiRlayer_trans.aiRlayer}
\title{aiRlayer_trans.aiRlayer}
\usage{
\method{aiRlayer_trans}{aiRlayer}(data, aiRlayer)
}
\arguments{
\item{data}{data of input values}

\item{aiRlayer}{takes either an aiRlayer, aiRnet, or aiRhidden object}
}
\value{
returns aiRnet output of data
}
\description{
aiRlayer_trans method for aiRlayer objects
}
