% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{aiRlayer_trans}
\alias{aiRlayer_trans}
\title{aiRlayer_trans}
\usage{
aiRlayer_trans(data, aiRlayer)
}
\arguments{
\item{data}{data of input values}

\item{aiRlayer}{takes either an aiRlayer, aiRnet, or aiRhidden object}
}
\value{
returns aiRnet output of data
}
