% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{aiRloss}
\alias{aiRloss}
\title{aiRloss}
\usage{
aiRloss(data, class.levels, classify)
}
\arguments{
\item{data}{aiRtransform data. data of activation nodes}

\item{class.levels}{character vector that identifies the
ordering of classification factor levels.}

\item{classify}{vector of correct training factors.}
}
\value{
returns aiRloss object which contains;
loss.prop: how far away each node was from desired activation,
row.loss: the total loss for an observation,
total.loss: the total loss for a batch of examples.
}
\description{
loss function used on data in training.
}
