% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{aiRnet}
\alias{aiRnet}
\title{aiRnet}
\usage{
aiRnet(nodes)
}
\arguments{
\item{nodes}{specifies how many nodes in each layer,
first value should equal input values, last value should
equal classification values}
}
\value{
generates aiRnet object of specified dimensions
}
\description{
Builds a network of aiRlayers objects with random weight and bias as a list
}
