% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{aiRrate}
\alias{aiRrate}
\title{aiRrate}
\usage{
aiRrate(data, factor, aiRnet, report.class = FALSE)
}
\arguments{
\item{data}{data of input values}

\item{factor}{character vector of factors to classify. levels must be in same order as
levels presented in original data.}

\item{aiRnet}{aiRnet object}

\item{report.class}{report results from aiRclassify}
}
\value{
error rate of aiRnet on data
}
\description{
reports error rate of network
}
