% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{aiRrowdelta}
\alias{aiRrowdelta}
\title{aiRrowdelta}
\usage{
aiRrowdelta(loss.prop, total.loss, aiRnet, n, data, index)
}
\arguments{
\item{loss.prop}{generated from aiRloss}

\item{total.loss}{generated from aiRloss}

\item{aiRnet}{aiRnet object}

\item{n}{length of aiRnet}

\item{data}{data used}

\item{index}{index of class vector}
}
\value{
new aiRnet object
}
\description{
back Propagation function
}
