% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{aiRrun}
\alias{aiRrun}
\title{aiRrun}
\usage{
aiRrun(data_train, data_test, aiRnet, classify_train.vec,
  classify_test.vec, class.levels, cycles, batch.size, modularTest)
}
\arguments{
\item{data_train}{training dataset}

\item{data_test}{testing dataset}

\item{aiRnet}{aiRnet object to be trained}

\item{classify_train.vec}{vector of training classifiers}

\item{classify_test.vec}{vector of test calssifiers}

\item{class.levels}{the levels of the classifier vector}

\item{cycles}{number of iterations done}

\item{batch.size}{number of samples taken at a time in a batch}

\item{modularTest}{How many cycles pass before the aiRnet is used on data_test}
}
\value{
list of aiRnet and dataframe
}
\description{
handles the activation and back propigation of the network.
}
