% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{aiRrun_test}
\alias{aiRrun_test}
\title{aiRrun_test}
\usage{
aiRrun_test(data, index, n, class.levels, sample.rows, aiRnet, cycles = 100,
  batch.size = "all")
}
\arguments{
\item{data}{Data frame that contains all named columns needed}

\item{index}{index of vector that contains classifying values}

\item{n}{number of layers in aiRnet}

\item{class.levels}{recorded classification levels}

\item{sample.rows}{logical vector generated by aiRsubset for data sampling}

\item{aiRnet}{aiRnet object generated by aiRnet function}

\item{cycles}{Number of cycles done to correct.}

\item{batch.size}{Number indicating how many rows to make batches from training sample. Default set to "all"
for no batches to be made.}
}
\value{
loss, aiRnet, traing.data and test.data
}
\description{
body of aiRrun. Includes test data subset and its loss calculations
}
