% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{aiRtransform}
\alias{aiRtransform}
\title{aiRtransform}
\usage{
aiRtransform(data, layers)
}
\arguments{
\item{data}{data of input values}

\item{layers}{aiRnet object}
}
\value{
returns layers output of data
}
