% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{approximate_change}
\alias{approximate_change}
\title{approximate_change}
\usage{
approximate_change(loss, aiRnet)
}
\arguments{
\item{loss}{aiRloss object created by aiRloss function}

\item{aiRnet}{aiRnet object created by aiRnet function}
}
\value{
return approximate change to aiRnet based on loss
}
\description{
approximates the change of an iterative gradient decent
}
