% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqi_category.R
\name{aqi_category}
\alias{aqi_category}
\alias{aqi_category_number}
\alias{aqi_category_color}
\title{Categorize API numbers}
\usage{
aqi_category(aqi)

aqi_category_number(aqi)

aqi_category_color(aqi)
}
\arguments{
\item{aqi}{A numeric Air Quality Index (AQI) value}
}
\value{
\code{aqi_category} returns a string with the AQI Category Descriptor

\code{aqi_category_number} returns a numeric value 1-6 indicating the corresponding AQI category number

\code{aqi_category_color} returns a string containing the hexidecimal color code associated with the corresponding AQI category number (e.g., \verb{#00E400} for "Good)
}
\description{
As described in \href{https://docs.airnowapi.org/aq101}{Air Quality 101}, the AQI is divided into a number of categories that indicate the potential health impacts of air conditions.
For example, a value of 23 corresponds to "Good", and is displayed with a green color.

\code{aqi_category} translates a given AQI number to it's category descriptor.

\code{aqi_category_number} TODO

\code{aqi_category_color} TODO
}
\examples{
aqi_category(23)
aqi_category_number(123)
aqi_category_color(23?a)
}
\seealso{
\href{https://docs.airnowapi.org/aq101}{Air Quality 101}
}
