% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aReal.R
\name{aReal}
\alias{aReal}
\alias{tWy}
\title{Areal interpolation}
\usage{
aReal(src, trg, vals)

tWy(W, y)
}
\arguments{
\item{src}{A \code{SpatialPolygonsDataFrame} from which to project.}

\item{trg}{A target geography on which to project values.}

\item{vals}{A vector containing the names of the columns in the \code{data}
slot of the \code{src} object to be used as source values.}
}
\value{
A \code{data.frame} containing the projected values along with
polygon IDs drawn from \code{trg}, along with information on the structure
of the common geographies within which values were interpolated.
}
\description{
\code{aReal} uses areal interpolation to project a set of source values on to 
a set of target observations.
}
\details{
Using the method proposed by Markoff and Shapiro (1973), \code{aReal} 
projects source values on to a target geography.  Source values are assumed
to be counts.
}
\examples{
#load list containing partitions
#data(nd_list)
#aReal(nd_list[[1]], nd_list[[2]], nd_list[[1]]$wheat)
}
\references{
Markoff, J. and Shapiro, G.  1973.  "The Linkage of Data
Describing Overlapping Geographical Units."  \emph{Historical Methods 
Newsletter} 7: 34--46.
}
