% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{backprop}
\alias{backprop}
\title{backprop}
\usage{
backprop(aiRnet, aiRcost, aiRactivation)
}
\arguments{
\item{aiRnet}{aiRnet object}

\item{aiRcost}{aiRcost object}

\item{aiRactivation}{aiRactivation object}
}
\value{
returns aiRnet with change.w and change.b values filled
}
\description{
backprop
}
