% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{cut2groups}
\alias{cut2groups}
\title{cut2groups}
\usage{
cut2groups(loss.prop)
}
\arguments{
\item{loss.prop}{loss.prop in aiRloss object}
}
\value{
summarised matrix of loss.prop
}
\description{
cuts loss.prop into groups based on how much each node should change
breaks are listed as c(-1.01,-.75,-.5,-.25,-.1,.1,.25,.5,.75,1)
}
