% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets_keyring.R
\name{delete_key}
\alias{delete_key}
\title{Deletes your securely stored OpenAI API key from your OS keyring.}
\usage{
delete_key()
}
\description{
Deletes your securely stored OpenAI API key from your OS keyring.
}
\examples{
delete_key()

}
