% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gateway_keyring.R
\name{delete_keyring_credentials}
\alias{delete_keyring_credentials}
\title{Deletes your securely stored OpenAI API key and preferred model
from your OS keyring.}
\usage{
delete_keyring_credentials()
}
\value{
\code{result} list of \code{success} or \code{failure} class with:
\item{status}{A character string of operation status.}
\item{value}{A character string of descriptive message of status.}
}
\description{
Deletes your securely stored OpenAI API key and preferred model
from your OS keyring.
}
\examples{
# CAREFUL: Deletes OpenAI API key and preferred model from your OS keyring
\dontrun{
delete_keyring_credentials()
}

}
