% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets_keyring.R
\name{delete_model}
\alias{delete_model}
\title{Deletes your stored OpenAI API model from your OS keyring.}
\usage{
delete_model()
}
\description{
Deletes your stored OpenAI API model from your OS keyring.
}
\examples{
delete_model()

}
