% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\docType{data}
\name{example}
\alias{example}
\title{Toy example from Slez, O'Connell, and Curtis (2014).}
\format{A data frame with 8 rows and 3 variables:
\itemize{
  \item A. A vector of numeric zone IDs associated with partition A.
  \item B. A vector of numeric zone IDs associated with partition B.
  \item EDGE.  A vector of numeric edge IDs.
}}
\usage{
data(example)
}
\description{
A dataset containing toy data on group overlap.
}
