% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{fresh.aiRlayer}
\alias{fresh.aiRlayer}
\title{fresh.aiRlayer}
\usage{
\method{fresh}{aiRlayer}(aiRlayer, rows = NULL, method.propigate = "fast")
}
\arguments{
\item{aiRlayer}{aiRlayer, aiRnet, or aiRhidden object}

\item{rows}{number of rows in a batch}
}
\value{
returns aiRnet object with change.w and change.b added
to weights and bias respectively
}
\description{
fresh method for aiRlayer objects
}
