% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_conditions.R
\name{get_conditions}
\alias{get_conditions}
\title{Get air quality conditions}
\usage{
get_conditions(
  zip = NULL,
  latitude = NULL,
  longitude = NULL,
  date = NULL,
  distance = 25,
  api_key = Sys.getenv("AIRNOW_API_KEY")
)
}
\arguments{
\item{zip}{US ZIP code}

\item{latitude}{TODO}

\item{longitude}{TODO}

\item{date}{Date at which to observe historical conditions. If not provided, current conditions will be returned.}

\item{distance}{If no data are associated with the given ZIP code, find observations from nearby reporting areas within this distance (in miles) (default: \code{25})}

\item{api_key}{AirNow API key. By default, the \code{AIRNOW_API_KEY} environment variable is used.}
}
\value{
Tibble containing current conditions for the given location
}
\description{
\lifecycle{experimental}

\code{get_conditions} gets forecast conditions for a given ZIP code or latitude/longitude pair
}
\note{
API keys can be created by signing up for an account at \url{https://docs.airnowapi.org}
}
\examples{
\dontrun{
get_conditions(zip = 90210)
}
}
