% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gateway_keyring.R
\name{get_credential}
\alias{get_credential}
\title{Gets a secret credential from your keyring or env variables.}
\usage{
get_credential(secret)
}
\description{
It first checks the environment variables, then the keyring.
}
