% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_conditions_zip.R
\name{get_current_conditions_zip}
\alias{get_current_conditions_zip}
\title{Get current conditions for a ZIP code}
\usage{
get_current_conditions_zip(
  zip,
  distance = 25,
  api_key = Sys.getenv("AIRNOW_API_KEY")
)
}
\arguments{
\item{zip}{US ZIP code}

\item{distance}{If no data are associated with the given ZIP code, find observations from nearby reporting areas within this distance (in miles) (default: \code{25})}

\item{api_key}{AirNow API key}
}
\value{
Tibble containing current conditions for the given ZIP code
}
\description{
Get current conditions for a ZIP code
}
