% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dcars.R
\name{get_dcars}
\alias{get_dcars}
\title{Pull data from both ASI and LERS}
\usage{
get_dcars(app_measures, enrol_measures, rows, institutions)
}
\arguments{
\item{app_measures}{Select possible measurements from ASI}

\item{enrol_measures}{Select possible measurements from LERS}

\item{institutions}{Filters by Institution}

\item{username}{Optional: Either supply a siams username or use .Rprofile otherwise as "siams.username"}

\item{password}{Optional: Either supply a siams password or use .Rprofile otherwise as "siams.password"}
}
\description{
Pull data from both ASI and LERS
}
\examples{
df <- get_dcars(app_measures = c("Unique Applicant Static"),
           enrol_measures = c("FLE"),
           rows = c("Provider", "Gender"),
           institutions = c("MH", "MU", "UA"))
}
