% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diversity.R
\name{get_diversity}
\alias{get_diversity}
\title{Pull data from LERS}
\usage{
get_diversity(rows, species, inst, type)
}
\arguments{
\item{rows}{Select rows to cut data by}

\item{type}{type of simpsondiversity index}
}
\description{
Pull data from LERS
}
