% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_enrolment.R
\name{get_enrolment}
\alias{get_enrolment}
\title{Pull data from LERS}
\usage{
get_enrolment(measures, rows, institutions, username, password, print = F,
  remove.offshores = T, remove.continuingstudies = T, postalcodes,
  sa.mh = F)
}
\arguments{
\item{measures}{Select possible measurements}

\item{rows}{Select rows to cut data by}

\item{institutions}{Optional: Filters by Institution}

\item{username}{Optional: Either supply a siams username or use .Rprofile otherwise as "siams.username".}

\item{password}{Optional: Either supply a siams password or use .Rprofile otherwise as "siams.password".}

\item{print}{Optional: Prints the MDX string used to query the database.}

\item{postalcodes}{Optional: Filters by Postal Code}
}
\description{
Pull data from LERS
}
\examples{
df <- get_enrolment(measures = c("Unique Student Static", "FLE"),
                rows = c("Provider", "Gender"),
                institutions = c("MH", "MU", "UA"))
}
