% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_forecast_zip.R
\name{get_forecast_zip}
\alias{get_forecast_zip}
\title{Get forecast for a ZIP code}
\usage{
get_forecast_zip(
  zip,
  date = Sys.Date(),
  distance = 25,
  api_key = Sys.getenv("AIRNOW_API_KEY")
)
}
\arguments{
\item{zip}{US ZIP code}

\item{date}{Date of forecast (default: current date)}

\item{distance}{If no data are associated with the given ZIP code, find observations from nearby reporting areas within this distance (in miles) (default: \code{25})}

\item{api_key}{AirNow API key}
}
\value{
Tibble containing forecasted conditions for the given ZIP code
}
\description{
Get forecast for a ZIP code
}
