% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets_keyring.R
\name{get_key}
\alias{get_key}
\title{Gets your OpenAI API key credential.}
\usage{
get_key()
}
\description{
Searches for your OpenAI API key in the following order:
1. Environment variable `OPENAI_KEY` – can be set in your `.Renviron` file
2. OS keyring – can be set securely with `set_key()`
}
