% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets_keyring.R
\name{get_model}
\alias{get_model}
\title{Gets your OpenAI API model preference.}
\usage{
get_model()
}
\description{
Searches for your OpenAI API model preference in the following order:
1. Environment variable `OPENAI_MODEL` – can be set in your `.Renviron` file
2. OS keyring – can be set securely with `set_model()`
}
