% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pfd.R
\name{get_pfd}
\alias{get_pfd}
\title{Pull Tidy Data from PFD}
\usage{
get_pfd(measures, rows, institutions, username, password, print = F)
}
\arguments{
\item{measures}{Select possible measurements}

\item{rows}{Optional: Select rows to cut data by}

\item{institutions}{Optional: Filters by Institution. Can exclude by including "exclude" in the vector, then the items in that vector will be excluded rather than included.}

\item{username}{Optional: Either supply a siams username or use .Rprofile otherwise as "siams.username".}

\item{password}{Optional: Either supply a siams password or use .Rprofile otherwise as "siams.password".}

\item{print}{Optional: Prints the MDX string used to query the database.}
}
\description{
Pull Tidy Data from PFD
}
\examples{
df <- get_pfd("Cost per FLE", "Program Band", "MH")
}
