% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_howto.R
\name{howto}
\alias{howto}
\title{Ask OpenAI to write R code for a problem.
The answer is written to console and returned as a string.}
\usage{
howto(do, call_api = call_openai)
}
\arguments{
\item{do}{Character string of what you want to do}

\item{call_api}{Optional function to call another API}
}
\value{
Invisible character string of R code
}
\description{
Ask OpenAI to write R code for a problem.
The answer is written to console and returned as a string.
}
\examples{
# You must provision an OpenAI API key before using this function.
\dontrun{
howto("read a csv file")
howto("get last element of a vector")
}

}
