% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_aqi.R
\name{is_aqi}
\alias{is_aqi}
\alias{is.aqi}
\title{Test AQI Values}
\usage{
is_aqi(x)
}
\arguments{
\item{x}{Object to be tested}
}
\value{
Logical value indicating whether the given value
}
\description{
\code{is_aqi} tests whether a given value is a valid AQI number or not.
}
\examples{
is_aqi(42)
is_aqi(-2)
}
\seealso{
https://docs.airnowapi.org/aq101
}
