% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lers_search.R
\name{lers_search}
\alias{lers_search}
\title{Pull Tidy Data for MHC from LERS using search string}
\usage{
lers_search(search_string, ...)
}
\arguments{
\item{search_string}{A string with a measure 'by' some row.}

\item{postalcodes}{Optional: Filters by Postal Code. Can exclude by including "exclude" in the vector, then the items in that vector will be excluded rather than included.}

\item{censusdivisions}{Optional: Filters by Census Division. Can exclude by including "exclude" in the vector, then the items in that vector will be excluded rather than included.}

\item{cipcodes}{Optional: Filters by Cip Code of any length, can used mixed vector with #'any length cip codes together. Can exclude by including "exclude" in the vector, then the items in that vector #'will be excluded rather than included.}

\item{ages}{Optional: Filters by age group.}

\item{username}{Optional: Either supply a siams username or use .Rprofile otherwise as "siams.username".}

\item{password}{Optional: Either supply a siams password or use .Rprofile otherwise as "siams.password".}

\item{print}{Optional: Prints the MDX string used to query the database.}
}
\description{
Pull Tidy Data for MHC from LERS using search string
}
\examples{
df <- ler_search("FLE by Gender", sa.mh = T, ages = 2)
}
