% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{mat.opperation}
\alias{mat.opperation}
\title{mat.opperation}
\usage{
mat.opperation(x, y, opperation)
}
\arguments{
\item{x}{matrix of column size m, row size n}

\item{y}{vector of length m}

\item{opperation}{opperation to use: +, -, *, /}
}
\value{
returns n by m matrix where values of y are done onto the columns of x.
}
\description{
allows simple operations between matrix and constant vector. column size of x must equal length of y
}
