% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mine_enrolment.R
\name{mine_enrolment}
\alias{mine_enrolment}
\title{Mine LERS data for permutations of rows and detect significant changes.}
\usage{
mine_enrolment(measure, rows, institutions = "MH", postalcodes = c(),
  censusdivisions = c(), cipcodes = c(), ages = c(), sa.mh = F,
  detect_change = F, min_year, max_year)
}
\arguments{
\item{rows}{Select elements in LERS to permutate combinations of.}

\item{institutions}{Optional: Supplys permutations for institutions.}

\item{postalcodes}{Optional: Filters by Postal Code. Can exclude by including "exclude" in the vector, then the items in that vector will be excluded rather than included.}

\item{censusdivisions}{Optional: Filters by Census Division. Can exclude by including "exclude" in the vector, then the items in that vector will be excluded rather than included.}

\item{cipcodes}{Optional: Filters by Cip Code of any length, can used mixed vector with any length cip codes together.}

\item{ages}{Optional: Filters by age group.}

\item{measures}{Select a measurement.}

\item{detect}{Use xmrr to detect significant changes.}
}
\description{
Mine LERS data for permutations of rows and detect significant changes.
}
\examples{
df <- mine_enrolment("FLE",
c("Program Type", "Credential Type", "Gender"),
institutions = c("MR", "MH", "UA"),
ages = 2, sa.mh = T)
}
