% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{narr_data}
\alias{narr_data}
\title{Daily Weather data.}
\format{A tibble with nine variables:
\describe{
\item{\code{vis}}{visibility}
\item{\code{hpbl}}{planetary boundary height (m)}
\item{\code{air.2m}}{air temperature at 2 m (K)}
\item{\code{rhum.2m}}{relative humidity at 2m}
\item{\code{prate}}{precipitation rate (kg/m^2/s)}
\item{\code{apcp}}{accumulated total precipitation (kg/m^2)}
\item{\code{pres.sfc}}{pressure (Pa)}
\item{\code{uwnd.10m}}{U-wind at 10 m (m/s)}
\item{\code{vwnd.10m}}{V-wind at 10 m (m/s)}
}}
\usage{
narr_data
}
\description{
Daily weather data for study area used within the PM2.5 prediction model.
Derived from the NCEP North American Regional Reanalysis (NARR).
}
\keyword{datasets}

