% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{neg.exp}
\alias{neg.exp}
\title{neg.exp}
\usage{
neg.exp(x, power = 2)
}
\arguments{
\item{x}{vector of numeric values}

\item{power}{value to exponentiate by. default set to 2.}
}
\value{
vector of numeric values equal in length to x
}
\description{
exponentiates x vector and retains negative values.
}
