% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing.R
\name{plot.aiR}
\alias{plot.aiR}
\title{plot aiR}
\usage{
\method{plot}{aiR}(aiR)
}
\arguments{
\item{aiRcost}{aiR object}
}
\value{
ggplot object
}
\description{
plotting the cost of the aiR object over all cycles
}
