% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing.R
\name{plot.aiRactivation}
\alias{plot.aiRactivation}
\title{plot aiRactivation}
\usage{
\method{plot}{aiRactivation}(aiRactivation, x.val, y.val, layer,
  nodes = NULL)
}
\arguments{
\item{aiRactivation}{aiRactivation object}

\item{x.val}{column name of value to plot on x axis}

\item{y.val}{column name of value to plot on y axis}

\item{layer}{name of which layer in aiRactivation to use}

\item{nodes}{non zero vector of integers that specify which nodes to view.
by default all nodes are plotted.}
}
\value{
ggplot object
}
\description{
plotting function for aiRactivation
}
