% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing.R
\name{plot.aiRnet}
\alias{plot.aiRnet}
\title{plot aiRnet}
\usage{
\method{plot}{aiRnet}(aiRnet, layers = NULL)
}
\arguments{
\item{aiRnet}{aiRnet object}

\item{layers}{non zero integer vector of layers to plot}
}
\value{
ggplot object
}
\description{
plots the weights and biases of each node in each layer of an aiRnet object
}
