% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pm_data_early}
\alias{pm_data_early}
\title{PM2.5 dataset (2000 - 2008).}
\format{A tibble with columns \code{pm_grid_id}, \code{date}, and
  \code{pm_pred}.}
\usage{
pm_data_early

pm_data_early
}
\description{
Dataset of estimated daily PM2.5 concentrations from 2000 - 2008. (as
microgram per meter cubed) corresponding to each \code{pm_grid_id} from the
\code{pm_grid}. The \code{pm_data} is split into two smaller datasets as a
workaround for GitHub's 100 MB filesize limit.  Please see the package README
for details.

Dataset of estimated daily PM2.5 concentrations from 2009 - 2015. (as
microgram per meter cubed) corresponding to each \code{pm_grid_id} from the
\code{pm_grid}. The \code{pm_data} is split into two smaller datasets as a
workaround for GitHub's 100 MB filesize limit.  Please see the package README
for details.
}
\keyword{datasets}

