% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pm_grid}
\alias{pm_grid}
\title{PM2.5 exposure assessment grid.}
\format{A simple features object continaing the coordinate geometry and
  corresponding \code{pm_grid_id}.}
\usage{
pm_grid
}
\description{
The 1 x 1 km grid used for exposure assessment.
}
\keyword{datasets}

