% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_result.R
\name{print.result}
\alias{print.result}
\title{make s3 print function for result, which prints its value}
\usage{
\method{print}{result}(x, ...)
}
\description{
make s3 print function for result, which prints its value
}
