% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{rowdelta.aiRnet}
\alias{rowdelta.aiRnet}
\title{rowdelta.aiRnet}
\usage{
\method{rowdelta}{aiRnet}(row.work, aiRlayer)
}
\arguments{
\item{row.work}{generated from aiRloss}

\item{aiRlayer}{either aiRlayer, aiRnet, or aiRhidden object}
}
\value{
new aiRnet object
}
\description{
rowdelta method for aiRnet objects

back Propagation function
}
