% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{rowdelta}
\alias{rowdelta}
\title{rowdelta}
\usage{
rowdelta(row.work, aiRlayer)
}
\arguments{
\item{row.work}{generated from aiRloss}

\item{aiRlayer}{either aiRlayer, aiRnet, or aiRhidden object}
}
\value{
new aiRnet object
}
\description{
back Propagation function
}
