% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gateway_keyring.R
\name{set_key}
\alias{set_key}
\title{Securely stores your your OpenAI API key to your OS keyring.}
\usage{
set_key(key = NULL)
}
\arguments{
\item{key}{Optional string of your OpenAI API key;
if not provided, a popup will ask you to enter it (safer).}
}
\description{
Securely stores your your OpenAI API key to your OS keyring.
}
\examples{
# CAREFUL: Changes your OpenAI API key in your OS keyring
\dontrun{
set_key("sk-my-api-key") # sets the key directly (will show in .Rhistory)
set_key()                # opens a system popup for secure entry
}

}
