% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_airnow.R
\name{use_airnow}
\alias{use_airnow}
\title{Register and Configure AirNow API Tokens}
\usage{
use_airnow(browse = rlang::is_interactive())
}
\arguments{
\item{browse}{Whether to open a browser window to create/view AirNow API token (default: \code{TRUE} for interactive sessions, \code{FALSE} otherwise)}
}
\description{
Register and Configure AirNow API Tokens
}
\examples{
\dontrun{use_airnow()}
}
