% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/utility.R
=======
% Please edit documentation in R/Functions.R
>>>>>>> f82d2399b6bb629cb94964c3a0632c5e76f9d69c
\name{zero}
\alias{zero}
\title{zero}
\usage{
<<<<<<< HEAD
zero(x, power = -5, base = 10, coef = 1)
=======
zero(x, power = -5, base = 10, coef = 1, replace = 0,
  inequal = "lesser")
>>>>>>> f82d2399b6bb629cb94964c3a0632c5e76f9d69c
}
\arguments{
\item{x}{numeric vector to turn to zero}

\item{power}{the power associated. default = -5}

\item{base}{Base of the exponent. default = 10}

\item{coef}{coefficent of the power and base. default = 1}
<<<<<<< HEAD
=======

\item{replace}{number to replace value by}

\item{inequal}{set to either lesser or greater, default = "lesser"}
>>>>>>> f82d2399b6bb629cb94964c3a0632c5e76f9d69c
}
\description{
Numbers under threshold are assigned numeric zero. Threshold default equals 0.00001
}
