% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip.R
\name{get_zip_conditions}
\alias{get_zip_conditions}
\alias{get_zip_forecast}
\title{Get air quality info for a US ZIP code}
\usage{
get_zip_conditions(zip, distance = 25, api_key = Sys.getenv("AIRNOW_API_KEY"))

get_zip_forecast(
  zip,
  date = Sys.Date(),
  distance = 25,
  api_key = Sys.getenv("AIRNOW_API_KEY")
)
}
\arguments{
\item{zip}{US ZIP code}

\item{distance}{If no data are associated with the given ZIP code, find observations from nearby reporting areas within this distance (in miles) (default: \code{25})}

\item{api_key}{AirNow API key}
}
\value{
Tibble containing current conditions for the given ZIP code
}
\description{
\code{get_zip_conditions} gets forecasted conditions for a given ZIP code

\code{get_zip_forecast} gets forecasted conditions for a given ZIP code
}
\details{
\lifecycle{experimental}
}
\examples{

}
