% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{aiRactivation}
\alias{aiRactivation}
\title{aiRactivation}
\usage{
aiRactivation(data, aiRnet, range.size = 1000)
}
\arguments{
\item{data}{data of input values}

\item{aiRnet}{aiRnet object}

\item{range.size}{total amount of data points used to visualize input space.}
}
\value{
returns list of data frames for each aiRlayer in aiRnet. Data frames show
activation level.
}
