% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organized.R
\name{aiRbatch}
\alias{aiRbatch}
\title{aiRbatch}
\usage{
aiRbatch(n, batch.size)
}
\arguments{
\item{batch.size}{how many observations are used per batch}

\item{data}{data of input values}
}
\value{
a list of data.frames containing random samples of input data.
}
\description{
randomly splits data into batches through sample. No repeated samples.
}
