% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{aiRfresh}
\alias{aiRfresh}
\title{aiRfresh}
\usage{
aiRfresh(aiRnet, rows, n)
}
\arguments{
\item{aiRnet}{aiRnet object}

\item{rows}{number of rows in a batch}

\item{n}{length of aiRnet object}
}
\value{
returns aiRnet object with change.w and change.b added
to weights and bias respectively
}
