% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{aiRrate}
\alias{aiRrate}
\title{aiRrate}
\usage{
aiRrate(data, factor, aiRnet, report.class = FALSE, warning = TRUE)
}
\arguments{
\item{data}{data of input values}

\item{factor}{character vector of factors to classify. levels must be in same order as
levels presented in original data.}

\item{aiRnet}{aiRnet object}

\item{report.class}{report results from aiRclassify, Default set to FALSE}

\item{warning}{Suppress warning from aiRclassify that multiple nodes are activated.
Default set to TRUE. TRUE: allow warning. FALSE: Suppress warning}
}
\value{
error rate of aiRnet on data and how many activated non uniquely.
}
\description{
reports error rate of network
}
