% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{aiRrowdelta}
\alias{aiRrowdelta}
\title{aiRrowdelta}
\usage{
aiRrowdelta(loss.prop, row.loss, total.loss, aiRnet, n)
}
\arguments{
\item{loss.prop}{generated from aiRloss}

\item{row.loss}{generated from aiRloss}

\item{total.loss}{generated from aiRloss}

\item{aiRnet}{aiRnet object}

\item{n}{length of aiRnet}
}
\value{
new aiRnet object
}
\description{
back Propagation function
}
