% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{aiRrun}
\alias{aiRrun}
\title{aiRrun}
\usage{
aiRrun(data, var.classify, aiRnet, cycles = 100, train.method = "Sample",
  sample.size = 0.5, batch.size = "all", train.Factor = NULL,
  test = FALSE, na.rm = TRUE, data.return = FALSE)
}
\arguments{
\item{data}{Data frame that contains all named columns needed}

\item{var.classify}{index or column name of vector that contains classifying values}

\item{aiRnet}{aiRnet object generated by aiRnet function}

\item{cycles}{Number of cycles done to correct.}

\item{train.method}{Method to save internal subset of data as the training data. "Sample" to take
a random sample of all rows in data as training set. "Factor" to indicate if you are training
on rows containing a particular factor level.}

\item{sample.size}{Number between (0-1) that modifies how much of desired train.method data is used.
Default set to 0.5. Excluded rows will be used as test examples and not affect aiRnet.}

\item{batch.size}{Number indicating how many rows to make batches from training sample. Default set to "all"
for no batches to be made.}

\item{train.Factor}{Necessary when train.method set to "Factor". Assign as vector of length 2 in the following form
c("column.index/name","factor.level"). Can assign only column name or index but first level is chosen in this case.}

\item{test}{toggle if TRUE if test set loss and error should be reported as well. Default set to FALSE}

\item{na.rm}{remove NAs, default set to TRUE. Function likely to fail with NAs}

\item{data.return}{logical determines if data should be returned to the user. Default set to FALSE}
}
\value{
data frame of loss and the aiRnet of the training values. data set used is optional
}
