% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{aiRsubset}
\alias{aiRsubset}
\title{aiRsubset}
\usage{
aiRsubset(x, train.method, sample.size, train.Factor = NULL)
}
\arguments{
\item{train.method}{Method to train data. "Sample" to take a sample of values in data. "Factor"}

\item{sample.size}{Number between (0-1) that modifies how much of training data is used.}

\item{train.Factor}{Necessary when train.method set to "Factor". Assign as vector of length 2 in the following form
c("column.index/name","factor.level"). Can assign only column name or index but first level is chosen in this case.}
}
\value{
Row numbers to subset
}
