% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{aiRtransform}
\alias{aiRtransform}
\title{aiRtransform}
\usage{
aiRtransform(data, aiRnet, n = NULL)
}
\arguments{
\item{data}{data of input values}

\item{aiRnet}{aiRnet object}
}
\value{
returns aiRnet output of data
}
