% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clustR.R
\name{clustR}
\alias{clustR}
\alias{mp_int}
\alias{mp_shp}
\title{Identify common geographies}
\usage{
clustR(plist, nid = NULL, area = NULL, thresh = 0.05)

mp_shp(plist, nid)

mp_int(plist, nid, area)
}
\arguments{
\item{plist}{Either a list of \code{SpatialPolygonsDataFrame} objects or a
list containing a single \code{data.frame} object.  If
\code{length(plist) > 1}, then \code{plist} should contain the original
paritions.  Otherwise, \code{plist} should contain a
\code{SpatialPolygonsDataFrame} or a \code{data.frame} depicting the
intersection between partitions.}

\item{nid}{A character vector containing the column names used to identify
groups within each partition.  This is only required when starting with
intersections as opposed to paritions.  When starting with partitions,
\code{clustR} will default to polygon IDs unless otherwise indicated through
the \code{nid} argument.  Groups should be uniquely identified
within partitions.}

\item{area}{A string containing the name of the column containing data on the
area of overlap between groups.  This is only required when using a single
\code{data.frame} object.}

\item{thresh}{A number between 0 and 1 used to drop ties resulting from
spurious polygons.  This value represents the area of group overlap expressed
as a proportion of the area of the smallest overlapping unit in question.}
}
\value{
A \code{data.frame} depicting the relationship between hyperedges and
components.  Each hyperedge consists of a membership profile containing the
name of one group from each partition.  Each component refers to a common
geography.
}
\description{
\code{clustR} identifies common geographies based on patterns of group overlap.
Input can take one of two forms: partitions or intersections.  A set of
partitions is represented by a list of \code{SpatialPolygonsDataFrame}
objects, each of which is composed of a set of areal units
(e.g., counties, census tracts).  To identify common geographies \code{clustR}
calculates the intersection of these partitions.  This can be very slow
depending on the number of observations and the detail of the underlying
boundary files.  Alternatively, intersections can be calculated using a
dedicated GIS (e.g., ArcGIS, QGIS) and then passesd to \code{clustR}
(recommended).  Intersections are represented as a list containing a single
object.  This can be either a \code{SpatialPolygonsDataFrame} object or a
\code{data.frame} object.  \code{mp_shp} and \code{mp_int} are internal helper
functions used to construct properly formatted membership profiles.
}
\details{
\code{clustR} assigns areal units to common geographies using the
method outlined by Slez, O'Connell, and Curtis (2014), who show that
identifying the common geographies associated with a set of \eqn{k} partitions
is identical to identifying the components of a \eqn{k}-uniform
\eqn{k}-partite hypergraph.  Each edge in the hypergraph represents a
membership profile depicting the intersection between areal units.
}
\examples{
#load list containing partitions
data(nd_list)
clustR(nd_list)

#load data frame containing intersections
data(example)

#add placeholder for area (real areas not needed)
example$AREA <- 1
clustR(list(example), nid = c("A", "B"), area = "AREA")

#load data frame containing intersections
data(south_df)
clustR(list(south_df), nid = c("ID1860", "ID2000"), area = "AREA")
}
\references{
Slez, A., O'Connell, H.A., and Curtis, K.J.  2014.  "A Note on the
Identification of Common Geographies."
}

