% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gateway_keyring.R
\name{delete_keyring_credentials}
\alias{delete_keyring_credentials}
\title{Deletes your securely stored OpenAI API key and preferred model
from your OS keyring.}
\usage{
delete_keyring_credentials()
}
\description{
Deletes your securely stored OpenAI API key and preferred model
from your OS keyring.
}
\examples{
delete_keyring_credentials()

}
