% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_applications.R
\name{get_applications}
\alias{get_applications}
\title{Pull data from ASI}
\usage{
get_applications(measures, rows, institutions, username, password, print = F,
  postalcodes, censusdivisions, sa.mh = F)
}
\arguments{
\item{measures}{Select possible measurements}

\item{rows}{Select rows to cut data by}

\item{institutions}{Optional: Filters by Institution}

\item{username}{Optional: Either supply a siams username or use .Rprofile otherwise as "siams.username"}

\item{password}{Optional: Either supply a siams password or use .Rprofile otherwise as "siams.password"}

\item{print}{Optional: Prints the MDX string used to query the database}

\item{postalcodes}{Optional: Filters by Postal Code}

\item{censusdivisons}{Optional: Filters by Census Division}
}
\description{
Pull data from ASI
}
\examples{
df <- get_applications(measures = c("Unique Applicant Static", "Application Record Count"),
                   rows = c("Provider", "Gender"),
                   institutions = c("MH", "MU", "UA"))
}
