% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diversity.R
\name{get_diversity}
\alias{get_diversity}
\title{Calculate a Simpson Diversity Index for LERS Data}
\usage{
get_diversity(species, rows, inst, type)
}
\arguments{
\item{species}{Element of diversity}

\item{rows}{Select rows to cut data by}

\item{inst}{Optional: Filters by Institution}

\item{type}{Optional: Type of Simpson Diversity Index, Defaults to "Dominance"}
}
\description{
Calculate a Simpson Diversity Index for LERS Data
}
