% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_enrolment.R
\name{get_enrolment}
\alias{get_enrolment}
\title{Pull Tidy Data from LERS}
\usage{
get_enrolment(measures, rows, institutions, username, password, print = F,
  remove.offshores = T, remove.continuingstudies = T, postalcodes,
  censusdivisions, cipcodes, sa.mh = F)
}
\arguments{
\item{measures}{Select possible measurements}

\item{rows}{Select rows to cut data by}

\item{institutions}{Optional: Filters by Institution. Can exclude by including "exclude" in the vector, then the items in that vector will be excluded rather than included.}

\item{username}{Optional: Either supply a siams username or use .Rprofile otherwise as "siams.username".}

\item{password}{Optional: Either supply a siams password or use .Rprofile otherwise as "siams.password".}

\item{print}{Optional: Prints the MDX string used to query the database.}

\item{postalcodes}{Optional: Filters by Postal Code. Can exclude by including "exclude" in the vector, then the items in that vector will be excluded rather than included.}

\item{censusdivisions}{Optional: Filters by Census Division. Can exclude by including "exclude" in the vector, then the items in that vector will be excluded rather than included.}

\item{cipcodes}{Optional: Filters by Cip Code of any length, can used mixed vector with any length cip codes together.}
}
\description{
Pull Tidy Data from LERS
}
\examples{
df <- get_enrolment(measures = c("Unique Student Static", "FLE"),
                rows = c("Provider", "Gender"),
                institutions = c("MH", "MU", "UA"),
                postalcodes = c("exclude", "T1C"),
                censusdivision = c("exclude", "1"))
}
