% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatis.R
\name{whatis}
\alias{whatis}
\title{Ask OpenAI to explain what given R code does.
The answer is written to console and returned as a string.}
\usage{
whatis(this, call_api = call_openai)
}
\arguments{
\item{this}{Character string of code to be explained}

\item{call_api}{Optional function to call another API}
}
\value{
Invisible character string of explanation
}
\description{
Ask OpenAI to explain what given R code does.
The answer is written to console and returned as a string.
}
\examples{
# howto("read a csv file")
# howto("get last element of a vector")

}
