% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{zero}
\alias{zero}
\title{zero}
\usage{
zero(x, power = -5, base = 10, coef = 1)
}
\arguments{
\item{x}{numeric vector to turn to zero}

\item{power}{the power associated. default = -5}

\item{base}{Base of the exponent. default = 10}

\item{coef}{coefficent of the power and base. default = 1}
}
\description{
Numbers under threshold are assigned numeric zero. Threshold default equals 0.00001
}
